+++ .gitignore
@@ -3,4 +3,3 @@
 *.idea
 **/debug/
 **/Cargo.lock
-**/target/
