+++ python/requirements.txt
@@ -1,4 +1,1 @@
 cryptography
-requests
-bs4
-lxml
