+++ rust/Cargo.toml
@@ -1,21 +1,8 @@
 [package]
-name = "crypto_pals"
+name = "rust"
 version = "0.1.0"
 authors = ["Neel Kuila <neelk1123@gmail.com>"]
-edition = "2018"
 
 [dependencies]
 base64 = "0.9.0"
 hex = "0.3.1"
-rand = "0.4.5"
-real_mt = {version="1.1.1", package="mersenne_twister"}
-chrono = "0.4.6"
-
-[[bin]]
-name = "s3c22"
-path = "src/s3c22.rs"
-
-[[bin]]
-name = "s3c23"
-path = "src/s3c23.rs"
-
